refactor: MyFont 컴포넌트 리팩토링 (#43)
