+++ src/shared/ui/SideNavigationBar.jsx
@@ -13,9 +13,9 @@ import {
 
 const NAV_LIST = [
   { label: '홈', Icon: NavHomeIcon, url: '/' },
+  { label: '둘러보기', Icon: NavExploreIcon, url: '/explore' },
   { label: '폰트 제작', Icon: NavFontDesignIcon, url: '/font-design' },
   { label: '폰트 합성', Icon: NavFontSynthesisIcon, url: '/font-synthesis' },
-  { label: '둘러보기', Icon: NavExploreIcon, url: '/explore' },
   { label: '북마크', Icon: NavBookmarkIcon, url: '/bookmark' },
   { label: '내 폰트', Icon: NavMyFontIcon, url: '/my-font' },
 ]
