+++ src/widgets/font-item/ui/FontItemWithProfile.jsx
@@ -5,7 +5,7 @@ import { BookmarkButton } from '@/features/bookmark/ui/BookmarkButton'
 import { DownloadButton } from '@/features/download/ui/DownloadButton'
 import { FontProfile } from '@/shared/ui/FontProfile'
 
-export const FontItemWithButtons = ({ font }) => (
+export const FontItemWithProfile = ({ font }) => (
   <LinkContainer to={`/detail/${font.id}`}>
     <FontProfile size="md" fontName={font.fontName} authorName={font.authorName} />
     <PreviewText>{font.preview}</PreviewText>
